webforms added, all forms are in there now for clarity
some visual changes

user can now only update himself, but can still delete others
