+++ templates/createuser.html
@@ -36,11 +36,14 @@
         {{ form.email.label(class="form-label") }}
         {{ form.email(class="form-control" ) }}
         <br>
+
         {{ form.password_hash.label(class="form-label") }}
         {{ form.password_hash(class="form-control" ) }}
+
         <br>
         {{ form.password_hash2.label(class="form-label") }}
         {{ form.password_hash2(class="form-control" ) }}
+
         <br>
         {{ form.address.label(class="form-label") }}
         {{ form.address(class="form-control" ) }}
