+++ templates/productpage.html
@@ -25,7 +25,10 @@
      <p><strong>Price:</strong> €{{'%.2f' % product['price']}}</p>
      <p><strong>Description:</strong> {{product['description']}} </p>
      <p><strong>This product was posted at: </strong>{{product['created_at']}} </p>
+     {% if current_user.is_authenticated %}
      <a class="btn btn-primary" href="{{ url_for('update_product', id=product.id)}}"> Edit </a>
+
+     {% endif %}
 
  </div>
 {% endblock %}
